Deployment/requirements.txt
numpy==1.23.5
scikit-learn==1.0.2
scikit-image==0.21.0
matplotlib==3.9.4
pandas==1.4.4
Pillow==10.4.0
streamlit==1.35.0
joblib==1.5.2
natsort==8.4.0
opencv-python-headless==4.5.5.64
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.2
unzip




